backend/requirements.txt
django
djangorestframework
requests
psycopg2
gunicorn
python-decouple
shortuuid